backend/requirements.txt
# ─── Web Framework ───────────────────────────────────────────
flask==3.0.3
flask-cors==4.0.1
flask-socketio==5.3.6
flask-limiter==3.7.0

# ─── Auth ────────────────────────────────────────────────────
PyJWT==2.8.0
bcrypt==4.1.3

# ─── Async ───────────────────────────────────────────────────
gevent==24.2.1

# ─── Database ────────────────────────────────────────────────
psycopg2-binary==2.9.9

redis==5.0.4
sqlalchemy==2.0.30

# ─── AI / LLM Clients ────────────────────────────────────────
groq==0.9.0
google-generativeai>=0.5.2,<0.6.0   # pinned to match langchain-google-genai 1.0.x
mistralai==0.4.2
openai==1.30.5                       # OpenRouter uses OpenAI-compatible API

# ─── LangGraph / LangChain ───────────────────────────────────
langgraph==0.1.5
langchain==0.2.5
langchain-community==0.2.5
langchain-groq==0.1.6
langchain-google-genai==1.0.6
langchain-openai==0.1.9

# ─── Embeddings & Vector Search ──────────────────────────────
sentence-transformers==3.0.1
pgvector==0.2.5
rank-bm25==0.2.2

# ─── Document Processing ─────────────────────────────────────
pdfplumber==0.11.0
pypdf2==3.0.1
python-multipart==0.0.9
supabase==2.5.0

# ─── Logging & Monitoring ────────────────────────────────────
structlog==24.2.0

# ─── Utilities ───────────────────────────────────────────────
python-dotenv==1.0.1
pydantic==2.7.3
httpx==0.27.0

# ─── Dev / Testing ────────────────────────────────────────────
pytest==8.2.2
pytest-mock==3.14.0
